@@ -31,11 +31,14 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@types/jest": "^29.4.0",
     "@types/testing-library__jest-dom": "^5.14.5",
+    "@typescript-eslint/eslint-plugin": "^5.55.0",
+    "@typescript-eslint/parser": "^5.55.0",
     "babel-jest": "^29.3.1",
     "css-loader": "^6.7.3",
     "cypress": "^12.3.0",
     "dotenv-webpack": "^8.0.1",
-    "eslint": "^8.35.0",
+    "eslint": "^8.36.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "html-webpack-plugin": "^5.5.0",
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.5.0",
